+++ CHANGELOG.md
@@ -15,12 +15,15 @@
 * Fixed an issue that prevented parsing tasks and functions with `inout` ports
 * Fixed conflicting genvar names when inlining interfaces and modules that use
   them; all genvars are now given a design-wide unique name
+* Fixed failure to resolve typenames suffixed with dimensions in contexts
+  permitting both types and expressions, e.g., `$bits(T[W-1:0])`
 * Fixed errant constant folding of shadowed non-trivial localparams
 * Fixed certain non-ANSI style port declarations being incorrectly reported as
   incompatible
 
 ### Other Enhancements
 
+* Added error checking for unresolved typenames
 * Added constant folding for `||` and `&&`
 
 ## v0.0.11